refactor: moved initial sprint screens into initial_dev folder
